abstracted configuration management for non-middleware use-cases
