+++ setup.py
@@ -5,7 +5,7 @@ from setuptools import setup, find_packages
 
 setup(
     name='arxiv-vault',
-    version='0.0.14',
+    version='0.0.15',
     packages=[f'arxiv.{package}' for package in find_packages('arxiv')],
     zip_safe=False,
     install_requires=[
